refactor: Refactored MatchController.cs and updated interfaces for structured JSON responses and data access.

+++ Interfaces/IMatchHistoryService.cs
@@ -2,6 +2,6 @@ namespace dotatryhard.Interfaces
 {
     public interface IMatchHistoryService
     {
-        Task<(List<int> Matches, List<int> Players)?> GetMatchHistoryAsync(int accountId);
+        Task<(List<long> Matches, List<int> Players)?> GetMatchHistoryAsync(int accountId);
     }
 }
